# docs/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Varka Backend API
  version: 0.0.1
  description: REST API for authentication, projects, CIM/PIM/PSM models, transformations, artifacts, modeling metadata, and diagram layout.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1:8080
    description: Local development
security:
  - AuthToken: []
...
paths:
...
  /api/transformations/cim-to-pim:
    post:
      tags: [Transformations]
      summary: Generate a PIM model from a CIM model
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/TransformRequest"
      responses:
        "202":
          $ref: "#/components/responses/MdeJobAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/transformations/pim-to-psm:
    post:
      tags: [Transformations]
      summary: Generate a PSM model from a PIM model
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/TransformRequest"
      responses:
        "202":
          $ref: "#/components/responses/MdeJobAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/transformations/psm-to-artifact:
    post:
      tags: [Transformations]
      summary: Generate deployable artifacts from a PSM model
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/TransformRequest"
      responses:
        "202":
          $ref: "#/components/responses/MdeJobAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
  /api/transformations/jobs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Transformations]
      summary: Get a transformation job
      responses:
        "200":
          description: Job record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MdeJobRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/transformations/jobs/{id}/cancel:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Transformations]
      summary: Cancel a transformation job
      responses:
        "200":
          description: Cancelled job record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MdeJobRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    AuthToken:
      type: apiKey
      in: header
      name: X-Auth-Token
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 256
      description: Replays the original async MDE job when the request fingerprint matches.
  requestBodies:
    TransformRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TransformRequest"
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Unauthorized:
      description: Missing, invalid, or expired auth session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Forbidden:
      description: Authenticated user does not have permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    NotFound:
      description: Record or file not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Conflict:
      description: Revision or state conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    PayloadTooLarge:
      description: Upload or generated payload exceeds configured limits
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
    MdeJobAccepted:
      description: MDE job accepted
      headers:
        Location:
          description: Polling URL for the accepted job
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JobResponse"
  schemas:
    ApiErrorResponse:
      type: object
      required: [message, status, timestamp, issues, errorId]
      properties:
        message:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        issues:
          type: array
          items: {}
        errorId:
          type: string
          description: Correlation id for support and log lookup
...
    ModelLevel:
      type: string
      enum: [CIM, PIM, PSM]
...
    ValidationResult:
      type: object
      required: [valid, issues]
      properties:
        valid:
          type: boolean
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
    ValidationIssue:
      type: object
      properties:
        severity:
          type: string
        constraint:
          type: string
        issueClass:
          type: string
        message:
          type: string
        guidance:
          type: string
        elementId:
          type: string
        elementName:
          type: string
...
    TransformRequest:
      type: object
      required: [sourceModelId]
      properties:
        sourceModelId:
          type: string
        expectedRevision:
          type: integer
          format: int64
...
    JobResponse:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED]
    MdeJobRecord:
      type: object
      required:
        [
          id,
          projectId,
          userId,
          sourceModelId,
          sourceLevel,
          sourceRevision,
          operation,
          status,
          progressPercent,
          diagnostics,
          createdAt,
        ]
      properties:
        id:
          type: string
        projectId:
          type: string
        userId:
          type: string
        sourceModelId:
          type: string
        sourceLevel:
          $ref: "#/components/schemas/ModelLevel"
        sourceRevision:
          type: integer
          format: int64
        sourceModelHash:
          type: string
        operation:
          type: string
          enum: [CIM_TO_PIM, PIM_TO_PSM, PSM_TO_ARTIFACT, VALIDATE]
        status:
          type: string
          enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED]
        progressPercent:
          type: integer
          minimum: 0
          maximum: 100
        resultModelId:
          type: string
        resultArtifactId:
          type: string
        diagnostics:
          type: array
          items:
            type: string
        validationResult:
          $ref: "#/components/schemas/ValidationResult"
        timings:
          type: object
          additionalProperties:
            type: integer
            format: int64
        idempotencyKey:
          type: string
        fingerprint:
          type: string
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
...